+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Description messages for pass and fail do not leave artifacts if there are no details
+- Make messages consistent with "Subject predicate noun"
 
 ## [0.4.0] - 2016-11-10
 ### Added