+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Fixed
+- [Fix false positive on value(included_in:) matcher](https://github.com/bloom-solutions/dry-validation-matchers/pull/10)
+
 ## [0.4.1] - 2018-06-01
 ### Fixed
 - [Fix date_time type error message](https://github.com/bloom-solutions/dry-validation-matchers/pull/9)